Moving stuff out of server.ts
